+++ client/src/ItemList.js
@@ -1,0 +1,36 @@
+import "./ItemListSmall.css";
+import ItemListContainer from "./components/ItemListContainer";
+
+import itemImg from "./images/item.png";
+
+function ItemList() {
+  function createDummyItems() {
+    const dummyItems = [];
+
+    for (let i = 0; i < 15; i++) {
+      dummyItems.push({
+        id: i + 1,
+        imgURL: itemImg,
+        description: {
+          name: `Infinity Gauntlet ${i + 1}`,
+          price: 9.99,
+          location: "Asgard, Asgard",
+        },
+      });
+    }
+
+    return dummyItems;
+  }
+
+  const dummyItems = createDummyItems();
+
+  console.log(dummyItems);
+
+  return (
+    <div>
+      <ItemListContainer items={dummyItems} />
+    </div>
+  );
+}
+
+export default ItemList;
